Added namespaces and autoloading
