+++ .gitignore
@@ -1,0 +1,2 @@
+example.php
+thepb-cli
