+++ README.md
@@ -1,2 +1,9 @@
-# thepb-sdk-php
+# ThePB
 An SDK for interacting with the TheP(aste)B.in API, using PHP.
+
+## Contributors
+@jdarmst
+@LouisT
+@ryancco
+
+TODO: Examples
