--- lib/Key.php
@@ -1,17 +1,0 @@
-<?php
-
-class Key
-{
-    final public static function validate($key)
-    {
-        $api      = new API;
-        $response = $api->post("/keys/validate", array(
-            "apikey" => $key,
-        ));
-
-        if ($response->valid == "true") {
-            return true;
-        }
-        return false;
-    }
-}
